+++ src/apps/seelenweg/modules/item/infra/UserApplication.tsx
@@ -1,16 +1,17 @@
 import { SeelenCommand, SeelenWegSide } from '@seelen-ui/lib';
-import { convertFileSrc, invoke } from '@tauri-apps/api/core';
+import { invoke } from '@tauri-apps/api/core';
 import moment from 'moment';
 import { memo, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
 import { BackgroundByLayersV2 } from '../../../components/BackgroundByLayers/infra';
-import { LAZY_CONSTANTS, updatePreviews } from '../../shared/utils/infra';
+import { updatePreviews } from '../../shared/utils/infra';
 
 import { Selectors } from '../../shared/store/app';
 import { parseCommand } from 'src/apps/shared/Command';
-import { useIcon, useWindowFocusChange } from 'src/apps/shared/hooks';
+import { FileIcon } from 'src/apps/shared/components/Icon';
+import { useWindowFocusChange } from 'src/apps/shared/hooks';
 
 import { PinnedWegItem, TemporalWegItem } from '../../shared/store/domain';
 
@@ -36,12 +37,6 @@ export const UserApplication = memo(({ item, onAssociatedViewOpenChanged }: Prop
   const devTools = useSelector(Selectors.devTools);
   const settings = useSelector(Selectors.settings);
   const focusedApp = useSelector(Selectors.focusedApp);
-
-  const iconSrc =
-    useIcon({
-      path: item.path,
-      umid: item.umid,
-    }) || convertFileSrc(LAZY_CONSTANTS.MISSING_ICON_PATH);
 
   const { t } = useTranslation();
   const calculatePlacement = (position: any) => {
@@ -91,7 +86,7 @@ export const UserApplication = memo(({ item, onAssociatedViewOpenChanged }: Prop
       className={cx({ 'associated-view-open': openPreview || openContextMenu })}
     >
       <WithContextMenu
-        items={getUserApplicationContextMenu(t, item, devTools, iconSrc) || []}
+        items={getUserApplicationContextMenu(t, item, devTools) || []}
         onOpenChange={(isOpen) => {
           setOpenContextMenu(isOpen);
           if (openPreview && isOpen) {
@@ -162,7 +157,7 @@ export const UserApplication = memo(({ item, onAssociatedViewOpenChanged }: Prop
             onContextMenu={(e) => e.stopPropagation()}
           >
             <BackgroundByLayersV2 prefix="item" />
-            <img className="weg-item-icon" src={iconSrc} />
+            <FileIcon className="weg-item-icon" path={item.path} umid={item.umid} />
             {notificationsCount > 0 && <div className="weg-item-badge">{notificationsCount}</div>}
             <div
               className={cx('weg-item-open-sign', {